Removed obsolete dependency script.module.requests.
This module was used during development of the addon, but was slowing down script execution since it could take upto a whole second to import the module. The code was therefore re-written to use the built-in python modules: urllib, urllib2 and httplib in preference, which greatly helped speed up the script execution.