Correction for scheduling future events immediately prior to a DST change due to a kodi bug.
Adjusted the epochFromLocalTimestamp function to calculate the epoch of an EPG timestamp string based on the current UTC offset. This will not take DST changes into account when scheduling a future event just prior to DST changes, since the kodi EPG is not DST aware (as of Krypton). This kodi bug is reported in the following thread: https://forum.kodi.tv/showthread.php?tid=323973

Therefore, we shall behave the same in order for future series links to work just prior to a DST change. As reported in the aforementioned thread, kodi will need to be restarted following the DST change in order to correct the times in the EPG. It's important to note that series links also won't work until the kodi restart has been performed. However, after reading the thread it's safe to say that a kodi restart is mandatory for PVR users following DST changes anyhow.